Added tests for writing one file in archive and short named files. TODO: Actually implement version for non-time-dependent comparison (now broken - times are related to current time)

+++ include/entry.h
@@ -26,7 +26,7 @@ public:
 #endif
     /*! A size of file
      */
-    long Size;
+    long long Size;
     /*! A byte offet for entry
      */
     int Offset;
